fix: critical bugs + code quality from senior review
Critical fixes:
- NR RAT misclassification: NR codes (0xB8xx+) were tagged as LTE
  because they share equipment_id 0x0B. Now checked before switch.
- Garbage cell keys for ML1 packets: extract_cell_key() now handles
  per-log-code layouts instead of assuming RRC OTA header everywhere.
- NR Cell ID truncation: CellPassport::cell_id widened to uint64_t
  to hold 36-bit NCI.
- NR-ARFCN truncation: parse_metadata reads uint32_t, not uint16_t.
- UB in bits(): guarded nbits==0 and nbits>=32 cases.
- TOCTOU in LinuxSource: start()/stop() use atomic exchange.
- Unbounded read_le: documented precondition, uses std::memcpy.

Quality fixes:
- Removed unused includes (<bit>, <iostream>)
- Added missing includes (<cstdio>, <array>, <cassert>, <cstring>)
- Removed pointless alignas(8) on QualcommPacketView
- Made to_channel_type constexpr noexcept
- CMake: target-specific warning suppression, Threads::Threads,
  CMAKE_CXX_EXTENSIONS OFF, Catch2 module path fix
- Style: designated initializers, anonymous namespace for print_cell,
  to_string(RatType) instead of static_cast<int>
- Fixed stale comments in main.cpp

+++ src/core/CellIdentity.h
@@ -37,7 +37,7 @@ struct GlobalCellKey {
   uint16_t mcc{0};
   uint16_t mnc{0};
   uint32_t tac{0};
-  uint32_t cell_id{0};
+  uint64_t cell_id{0};
   auto operator<=>(const GlobalCellKey&) const = default;
 };
 
@@ -60,7 +60,7 @@ struct LocalCellKey {
 ///   - Network analysis: q_rx_lev_min, intra_freq_reselection_allowed
 struct CellPassport {
   uint32_t tac{0};      ///< Tracking Area Code (16-bit LTE / 24-bit NR)
-  uint32_t cell_id{0};  ///< Cell Identity (28-bit ECI for LTE, 36-bit NCI for NR)
+  uint64_t cell_id{0};  ///< Cell Identity (28-bit ECI for LTE, 36-bit NCI for NR)
   uint16_t mcc{0};      ///< Mobile Country Code (e.g. 250 = Russia)
   uint16_t mnc{0};      ///< Mobile Network Code (e.g. 01 = MTS)
